+++ frontend/src/app/game-board/game-board.css
@@ -15,14 +15,15 @@
 .tank {
   position: fixed;
   z-index: 2;
-  width: 40;
-  height: 64;
+  width: 40px;
+  height: 64px;
 }
 
 .wall {
   position: fixed;
   z-index: 3;
   object-fit: scale-down;
+  background-size: 48px;
   padding: 0;
   margin: 0;
 }
