+++ gamelogic/netcode/client.d.ts
@@ -5,4 +5,4 @@ export function shootBullet (x: number, y: number): void;
 export function join (cb: () => void): void;
 export function fetchFrame () : GameState | undefined;
 export function getDistance(idx1: number) : number | undefined;
-export function getClientIdx(): number | undefined;
+export function getClientIdx(): number;
