+++ gamelogic/netcode/client.js
@@ -1,8 +1,9 @@
 import { WebSocketService } from "./web-socket-service";
 import { initialize, step, shoot, move } from "../gamelogic/game-state";
-import { maxStepSize, serverStepSize } from "./common";
+import { maxStepSize, serverStepSize, isDef } from "./common";
 
 const wss = new WebSocketService();
+wss.setDelay(0);
 initialize();
 
 let clientInfo;
@@ -10,9 +11,9 @@ let currentState;
 let serverStates;
 let unprocessedInputs;
 
-
 export function moveTo (x, y) {
   if (clientInfo) {
+    console.log(x, y, clientInfo);
     wss.emit("game.move", { x, y, gameId: clientInfo.gameId, clientIdx: clientInfo.clientIdx });
     unprocessedInputs.push({
       timestamp: Date.now() - serverStepSize,
@@ -35,11 +36,7 @@ export function shootBullet (x, y) {
   }
 }
 
-export function join (cb ) {
-  wss.emit("match.joinRequest", { }, (c) => {
-    clientInfo = c;
-  });
-
+export function join (cb) {
   wss.bindHandler("match.join", (match) => {
     cb();
 
@@ -51,7 +48,11 @@ export function join (cb ) {
     wss.bindHandler("match.stateUpdate", (res) => {
       serverStates = res.newStates.concat(serverStates).sort((a, b) => a.timestamp - b.timestamp);
     })
-  })
+  });
+
+  wss.emit("match.joinRequest", { }, (c) => {
+    clientInfo = c;
+  });
 }
 
 export function fetchFrame () {
@@ -62,7 +63,10 @@ export function fetchFrame () {
   // then that input has already been processed
 
   const targetTime = Date.now();
-  for (let i = serverStates.length; i >= 0; i--) {
+  if (serverStates.length !== 0)
+    console.log(serverStates);
+
+  for (let i = serverStates.length - 1; i >= 0; i--) {
     if (currentState.timestamp <= serverStates[i].timestamp && serverStates[i].timestamp <= targetTime) {
       currentState = serverStates[i];
       break;
@@ -86,7 +90,7 @@ export function fetchFrame () {
     let delta = 0;
 
     const input = unprocessedInputs[0];
-    console.assert(!input || ((headTime < input.timestamp && input.timestamp < targetTime)));
+    console.assert(!isDef(input) || ((headTime < input.timestamp && input.timestamp < targetTime)), "non-null input w/ incorrect timestamp");
 
     delta = Math.min(maxStepSize, targetTime - headTime);
 
@@ -107,7 +111,7 @@ export function fetchFrame () {
       }
     }
 
-    console.assert(delta > 0);
+    console.assert(delta > 0, "negative delta");
 
     step(currentState, delta);
     headTime += delta;
@@ -120,10 +124,8 @@ export function getDistance(idx) {
   const t1 = currentState.tanks[idx];
   if (!t1 || !currentState) return;
 
-  assert(currentState);
-
   const t2 = currentState.tanks.clientInfo[clientInfo.clientIdx];
-  assert(t2);
+  assert(isDef(t2), "null client tank");
 
   return Math.sqrt(
     (t1.sprite.x - t2.sprite.x) ** 2 + (t1.sprite.y - t2.sprite.y) ** 2
@@ -131,5 +133,6 @@ export function getDistance(idx) {
 }
 
 export function getClientIdx() {
-  return clientInfo ? clientInfo.clientIdx : undefined;
+  console.assert(isDef(clientInfo), "null clientInfo");
+  return clientInfo.clientIdx;
 }
