+++ gamelogic/netcode/common.js
@@ -1,2 +1,6 @@
 export const serverStepSize = 100;
 export const maxStepSize = 20;
+
+export function isDef(x) {
+  return x !== undefined && x !== null;
+}
