+++ gamelogic/package.json
@@ -11,6 +11,7 @@
   "description": "",
   "dependencies": {
     "dotenv": "^17.0.0",
+    "express": "^5.1.0",
     "express-session": "^1.18.1",
     "socket.io": "^4.8.1",
     "socket.io-client": "^4.8.1"
